+++ packages/multichain-account-service/src/providers/EvmAccountProvider.test.ts
@@ -1,5 +1,9 @@
 import type { Messenger } from '@metamask/base-controller';
-import type { InternalAccount } from '@metamask/keyring-internal-api';
+import type { KeyringMetadata } from '@metamask/keyring-controller';
+import type {
+  EthKeyring,
+  InternalAccount,
+} from '@metamask/keyring-internal-api';
 
 import { EvmAccountProvider } from './EvmAccountProvider';
 import {
@@ -7,6 +11,8 @@ import {
   getRootMessenger,
   MOCK_HD_ACCOUNT_1,
   MOCK_HD_ACCOUNT_2,
+  MOCK_HD_KEYRING_1,
+  MockAccountBuilder,
 } from '../tests';
 import type {
   AllowedActions,
@@ -14,6 +20,51 @@ import type {
   MultichainAccountServiceActions,
   MultichainAccountServiceEvents,
 } from '../types';
+
+class MockEthKeyring implements EthKeyring {
+  readonly type = 'MockEthKeyring';
+
+  readonly metadata: KeyringMetadata = {
+    id: 'mock-eth-keyring-id',
+    name: '',
+  };
+
+  readonly accounts: InternalAccount[];
+
+  constructor(accounts: InternalAccount[]) {
+    this.accounts = accounts;
+  }
+
+  async serialize() {
+    return 'serialized';
+  }
+
+  async deserialize(_: string) {
+    // Not required.
+  }
+
+  getAccounts = jest
+    .fn()
+    .mockImplementation(() => this.accounts.map((account) => account.address));
+
+  addAccounts = jest.fn().mockImplementation((numberOfAccounts: number) => {
+    const newAccountsIndex = this.accounts.length;
+
+    // Just generate a new address by appending the number of accounts owned by that fake keyring.
+    for (let i = 0; i < numberOfAccounts; i++) {
+      this.accounts.push(
+        MockAccountBuilder.from(MOCK_HD_ACCOUNT_1)
+          .withUuid()
+          .withAddressSuffix(`${this.accounts.length}`)
+          .get(),
+      );
+    }
+
+    return this.accounts
+      .slice(newAccountsIndex)
+      .map((account) => account.address);
+  });
+}
 
 /**
  * Sets up a EvmAccountProvider for testing.
@@ -38,10 +89,31 @@ function setup({
     MultichainAccountServiceActions | AllowedActions,
     MultichainAccountServiceEvents | AllowedEvents
   >;
+  keyring: MockEthKeyring;
+  mocks: {
+    getAccountByAddress: jest.Mock;
+  };
 } {
+  const keyring = new MockEthKeyring(accounts);
+
   messenger.registerActionHandler(
     'AccountsController:listMultichainAccounts',
     () => accounts,
+  );
+
+  const mockGetAccountByAddress = jest
+    .fn()
+    .mockImplementation((address: string) =>
+      keyring.accounts.find((account) => account.address === address),
+    );
+  messenger.registerActionHandler(
+    'AccountsController:getAccountByAddress',
+    mockGetAccountByAddress,
+  );
+
+  messenger.registerActionHandler(
+    'KeyringController:withKeyring',
+    async (_, operation) => operation({ keyring, metadata: keyring.metadata }),
   );
 
   const provider = new EvmAccountProvider(
@@ -51,6 +123,10 @@ function setup({
   return {
     provider,
     messenger,
+    keyring,
+    mocks: {
+      getAccountByAddress: mockGetAccountByAddress,
+    },
   };
 }
 
@@ -83,5 +159,81 @@ describe('EvmAccountProvider', () => {
     expect(() => provider.getAccount(unknownAccount.id)).toThrow(
       `Unable to find account: ${unknownAccount.id}`,
     );
+  });
+
+  it('does not re-create accounts (idempotent)', async () => {
+    const accounts = [MOCK_HD_ACCOUNT_1, MOCK_HD_ACCOUNT_2];
+    const { provider } = setup({
+      accounts,
+    });
+
+    const newAccounts = await provider.createAccounts({
+      entropySource: MOCK_HD_KEYRING_1.metadata.id,
+      groupIndex: 0,
+    });
+    expect(newAccounts).toHaveLength(1);
+    expect(newAccounts[0]).toStrictEqual(MOCK_HD_ACCOUNT_1);
+  });
+
+  it('throws if the created account is not BIP-44 compatible', async () => {
+    const accounts = [MOCK_HD_ACCOUNT_1];
+    const { provider, mocks } = setup({
+      accounts,
+    });
+
+    mocks.getAccountByAddress.mockReturnValue({
+      ...MOCK_HD_ACCOUNT_1,
+      options: {}, // No options, so it cannot be BIP-44 compatible.
+    });
+
+    await expect(
+      provider.createAccounts({
+        entropySource: MOCK_HD_KEYRING_1.metadata.id,
+        groupIndex: 0,
+      }),
+    ).rejects.toThrow('Created account is not BIP-44 compatible');
+  });
+
+  it('throws when trying to create gaps', async () => {
+    const { provider } = setup({
+      accounts: [MOCK_HD_ACCOUNT_1],
+    });
+
+    await expect(
+      provider.createAccounts({
+        entropySource: MOCK_HD_KEYRING_1.metadata.id,
+        groupIndex: 10,
+      }),
+    ).rejects.toThrow('Trying to create too many accounts');
+  });
+
+  it('throws if internal account cannot be found', async () => {
+    const { provider, mocks } = setup({
+      accounts: [MOCK_HD_ACCOUNT_1],
+    });
+
+    // Simulate an account not found.
+    mocks.getAccountByAddress.mockImplementation(() => undefined);
+
+    await expect(
+      provider.createAccounts({
+        entropySource: MOCK_HD_KEYRING_1.metadata.id,
+        groupIndex: 1,
+      }),
+    ).rejects.toThrow('Internal account does not exist');
+  });
+
+  it('discover accounts', async () => {
+    const { provider } = setup({
+      accounts: [], // No accounts by defaults, so we can discover them
+    });
+
+    // TODO: Update this once we really implement the account discovery.
+    expect(
+      await provider.discoverAndCreateAccounts({
+        entropySource: MOCK_HD_KEYRING_1.metadata.id,
+        groupIndex: 0,
+      }),
+    ).toStrictEqual([]);
   });
 });