+++ packages/transaction-controller/src/TransactionController.test.ts
@@ -1511,6 +1511,10 @@ describe('TransactionController', () => {
           to: ACCOUNT_MOCK,
         },
         {
+          assetsFiatValues: {
+            sending: '100',
+            receiving: '50',
+          },
           deviceConfirmedOn: mockDeviceConfirmedOn,
           origin: mockOrigin,
           securityAlertResponse: mockSecurityAlertResponse,
@@ -1525,6 +1529,10 @@ describe('TransactionController', () => {
 
       expect(updateSwapsTransactionMock).toHaveBeenCalledTimes(1);
       expect(transactionMeta.txParams.from).toBe(ACCOUNT_MOCK);
+      expect(transactionMeta.assetsFiatValues).toStrictEqual({
+        sending: '100',
+        receiving: '50',
+      });
       expect(transactionMeta.chainId).toBe(MOCK_NETWORK.chainId);
       expect(transactionMeta.deviceConfirmedOn).toBe(mockDeviceConfirmedOn);
       expect(transactionMeta.origin).toBe(mockOrigin);
@@ -1717,6 +1725,7 @@ describe('TransactionController', () => {
 
       const expectedInitialSnapshot = {
         actionId: undefined,
+        assetsFiatValues: undefined,
         batchId: undefined,
         chainId: expect.any(String),
         dappSuggestedGasFees: undefined,