+++ packages/profile-sync-controller/CHANGELOG.md
@@ -7,6 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `UserStorageController` optional config callback `getIsMultichainAccountSyncingEnabled`, and `getIsMultichainAccountSyncingEnabled` public method / messenger action
+  - This callback needs to be wired to client specific selectors in order to fetch the value of the feature flag dynamically
+  - If `true`, Account syncing will stop pushing new data to the user storage and only act as an account restoration method that will be fired before multichain account syncing for legacy compatibility
+  - This is done because `AccountTreeController` will become responsible for Multichain Account syncing
+
+### Changed
+
+- Bump `@noble/hashes` from `^1.4.0` to `^1.8.0` ([#6101](https://github.com/MetaMask/core/pull/6101))
+- Bump `@noble/ciphers` from `^0.5.2` to `^1.3.0` ([#6101](https://github.com/MetaMask/core/pull/6101))
+
+### Removed
+
+- Unused `UserStorageController:saveInternalAccountToUserStorage` public method
+
 ## [23.0.0]
 
 ### Changed